fix fails because of asset precompilation and digests in 4.2
the block for Combustion.initialize! sets the old values for the test
env that turn off digests, so the matcher for the image tag in "should
be able to use bip_bool to change a boolean value using an image" works.

also moving css files to standard application.css manifest avoids the
complaints from rails about "Asset was not declared to be precompiled in
production." for each of these.